requirements.txt
fastapi
uvicorn
google-cloud-vision
firebase-admin
python-multipart
python-dotenv
